% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-winScan.R
\docType{methods}
\name{winScan}
\alias{.winSlider}
\alias{.winStats}
\alias{winScan}
\alias{winScan,data.frame-method}
\title{Apply several functions in a sliding window}
\usage{
winScan(x, ...)

\S4method{winScan}{data.frame}(x, groups = NULL, position = NULL, values,
  win_size, win_step = 0.5 * win_size, funs = "mean", cores = 1)

.winSlider(x, values, win_size, win_step, funs, cores)

.winStats(win_start, win_size, funs, values, x)
}
\arguments{
\item{x}{a data.frame object on which to calculate window statistics.}

\item{...}{other arguments passed to specific methods.}

\item{groups}{vector with the name of variables on which to group the data.frame.
It can be set to NULL, if there are no groups. Default: NULL}

\item{position}{name of variable which contains the positions on which to calculate the windows.
If set to NULL, a "rolling" window will be computed instead. Default: NULL}

\item{values}{vector of variables on which to calculate statistics.}

\item{win_size}{size of the window.}

\item{win_step}{step of the window. Default: 0.5*win_size.}

\item{funs}{vector of functions to compute for each window. Default: "mean".}

\item{cores}{number of processing cores.}

\item{x}{data.frame on which to calculate window statistics.}
}
\value{
a data.frame of window-computed statistics.
}
\description{
\code{winScan()} applies a list of functions to each of a list of variables, using
a sliding window approach. This can be a "rolling" window or a "position" window (see details).
If required, the windows can be defined independently for one or more grouping variables.

\code{.winSlider()} is an internal function called by \code{winScan()}.
It creates sliding window coordinates and applies a list of
functions to each window.

\code{.winStats()} is an internal function called by \code{mclapply()}
in \code{.winSlide()}.
}
\details{
A "rolling" window is based on consecutive rows in the \code{data.frame}. For example, if
`win_size = 10` and `win_step = 5`, then, on a data.frame with 100 rows there will be
20 windows, each containing 10 rows of the data.frame.
A "position" window is based on a variable containing positions.
}
\examples{
... to be filled in
... to be added
}

